@@ -12,8 +12,11 @@
   ],
   "minimum-stability": "stable",
   "require": {
-    "php": ">=7.0",
+    "php": ">=7.1",
     "chilimatic/interfaces-component": "2.*"
+  },
+  "require-dev": {
+    "phpunit/phpunit": "6.5.*"
   },
   "autoload": {
     "psr-4": {
