@@ -1,11 +1,5 @@
 <?php
-/**
- * Created by PhpStorm.
- * User: j
- * Date: 28.10.14
- * Time: 11:44
- */
-
+declare(strict_types=1);
 namespace chilimatic\lib\Cache\Engine;
 
 /**
@@ -21,7 +15,7 @@ trait CacheTrait
      *
      * @var array
      */
-    protected $cacheListing = array();
+    private $entryMetaData = [];
 
     /**
      * md5 sum of the cache_list from
@@ -56,9 +50,57 @@ trait CacheTrait
     /**
      * @param bool $connected
      */
-    protected function setConnected(bool $connected = true)
+    protected function setConnected(bool $connected = true): void
     {
-        $this->connected = (bool)$connected;
+        $this->connected = $connected;
+    }
+
+    /**
+     * @param string $key
+     * @param int $expiration
+     */
+    private function updateMetaData(string $key, int $expiration): void
+    {
+        $expiration = empty($expiration) ? 0 : $expiration;
+
+        $newListing = [
+            'key'        => $key,
+            'expiration' => $expiration,
+            'updated'    => (string)date('Y-m-d H:i:s')];
+
+        if (isset($this->entryMetaData[$key])) {
+            $this->entryMetaData[$key] = array_merge($this->entryMetaData[$key], $newListing);
+        } else {
+            $newListing['created']    = $newListing ['updated'];
+            $this->entryMetaData[$key] = $newListing;
+        }
+    }
+
+    /**
+     * delete cached values based on an input array
+     *
+     * @param array $keyArray
+     * @return bool
+     */
+    public function deleteFromList(array $keyArray = []) : bool
+    {
+        if (!$keyArray) {
+            return false;
+        }
+
+        foreach ($keyArray as $key_del) {
+            if (!$this->entryMetaData) {
+                break;
+            }
+
+            foreach ($this->entryMetaData as $key) {
+                if (stripos($key, strtolower($key_del) !== false)) {
+                    $this->delete($key);
+                }
+            }
+        }
+
+        return true;
     }
 
 }
