@@ -1,12 +1,5 @@
 <?php
-/**
- *
- * @author j
- * Date: 3/20/15
- * Time: 4:39 PM
- *
- * File: Model.php
- */
+declare(strict_types=1);
 
 namespace chilimatic\lib\cache\handler;
 
